Changes following review

+++ legend-engine-core/legend-engine-core-base/legend-engine-core-executionPlan-execution/legend-engine-executionPlan-dependencies/src/main/java/org/finos/legend/engine/plan/dependencies/util/Library.java
@@ -1784,7 +1784,7 @@ public class Library
         return new JaroWinklerSimilarity().apply(str1, str2);
     }
 
-    public static int levenshteinDistance(String str1, String str2)
+    public static long levenshteinDistance(String str1, String str2)
     {
         return new LevenshteinDistance().apply(str1, str2);
     }
